nfstest: add makefile target to show test results
Under the hood, it more or less just does 'find ... | xargs cat'.

It shows the text results files for the most recent kernel run.  You can
show the results for a different kernel by overriding the LAST_KERNEL
variable, e.g.

$ LAST_KERNEL=6.13.4-300.fc41.x86_64 make nfstest-show-results

Note that since the nfstest log files have a date and timestamp in the
filename, it's possible to accumulate logs from multiple test runs.  If
you're showing the results from the most recent kernel, then we'll only
show the logs from the most recent run (i.e. logs that are newer than
the timestamp of the last-run directory). But if you're showing the
results from an older kernel, then we'll show all the logs in that
directory (but you can change that by overriding the PATTERN variable).

Signed-off-by: Scott Mayhew <smayhew@redhat.com>